+++ src/utils/type.ts
@@ -1,0 +1,3 @@
+import React from "react";
+
+export type SetState<T> = React.Dispatch<React.SetStateAction<T>>;
